% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scounts-package.R
\docType{package}
\name{scounts-package}
\alias{scounts}
\alias{scounts-package}
\title{isomiRs}
\description{
Sequencing bias correction and differential expression.
}
\author{
\strong{Maintainer}: Lorena Pantano \email{lorena.pantano@gmail.com}

Authors:
\itemize{
  \item Christos Argyropoulos \email{cargyropoulos@unm.edu}
}

}
